Extractors

MCGraphs.Extractors — Module

Extract data from a metacells Daf for standard graphs.

Data

MCGraphs.Extractors.extract_gene_gene_data — Function

extract_gene_gene_data(
    daf::DafReader;
    x_gene::AbstractString,
    y_gene::AbstractString,
    axis::QueryString,
    [min_significant_gene_UMIs::Integer = 40,
    color::Maybe{QueryString} = nothing,
    entries_hovers::Maybe{FrameColumns} = ["total_UMIs" => "=", "type" => "="],
    genes_hovers::Maybe{FrameColumns} = nothing],
)::PointsGraphData

Extract the data for a gene-gene graph from the daf data. The X coordinate of each point is the fraction of the x_gene and the Y coordinate of each gene is the fraction of the y_gene in each of the entries of the axis . By default we show the genes fractions of all metacells.

We ignore genes that don't have at least min_significant_gene_UMIs between both entries.

If a colors_query is specified, it can be a suffix of a query that fetches a value for each entry of the axis , or a full query that groups by the axis. By default we color by the type of the axis .

Anything listed in hovers will be collected per metacell and used in the hover data.

MCGraphs.Extractors.extract_box_box_data — Function

function extract_box_box_data(
    daf::DafReader;
    x_box::AbstractString,
    y_box::AbstractString,
    [min_significant_gene_UMIs::Integer = 40,
    max_box_span::AbstractFloat = 2.0,
    gene_fraction_regularization::AbstractFloat = 1e-5],
)::PointsGraphData

Extract the data for a box-box graph. This shows why two boxes were not merged (or, if given the same box name twice, why the box was merged).

MCGraphs.Extractors.extract_marker_genes_data — Function

function extract_marker_genes_data(
    daf::DafReader;
    [axis::AbstractString = "metacell",
    min_marker_genes::Integer = 100,
    type_property::Maybe{AbstractString} = "type",
    expression_annotations::Maybe{FrameColumns} = nothing,
    gene_annotations::Maybe{FrameColumns} = ["is_lateral", "divergence"],
    gene_names::Maybe{Vector{AbstractString}} = nothing],
)::HeatmapGraphData

Extract the data for a marker genes graph. This shows the genes that most distinguish between metacells (or profiles using another axis). Type annotations are added based on the type_property (which should name an axis with a color property), and optional expression_annotations and gene_annotations .

If gene_names is specified, these genes will always appear in the graph. This list is supplemented with additional is_marker genes to show at least min_marker_genes . A number of strongest such genes is chosen from each profile, such that the total number of these genes together with the forced gene_names is at least min_marker_genes .

Configuration

MCGraphs.Extractors.extract_categorical_color_palette — Function

extract_categorical_color_palette(
    daf::DafReader,
    query::QueryString
)::AbstractVector{<:AbstractString, <:AbstractString}

Convert the results of a query to a color palette for rendering. The query must return a named vector of strings; the names are taken to be the values and the strings are taken to be the color names. For example, extract_categorical_color_palette(daf, q"/ type : color") will return a color palette mapping each type to its color.

MCGraphs.Extractors.default_gene_gene_configuration — Function

default_gene_gene_configuration(
    configuration = PointsGraphConfiguration();
    [gene_fraction_regularization::Maybe{AbstractFloat} = 1e-5],
)::PointsGraphConfiguration

Return a default configuration for a gene-gene graph. This just applies the log scale to the axes. Will modify configuration in-place and return it.

MCGraphs.Extractors.default_box_box_configuration — Function

default_box_box_configuration(
    [configuration::PointsGraphConfiguration = PointsGraphConfiguration()];
    x_box::AbstractString,
    y_box::AbstractString,
    x_neighborhood::AbstractString,
    [max_box_span::AbstractFloat = 2.0,
    gene_fraction_regularization::AbstractFloat = 1e-5],
)::PointsGraphConfiguration

Return a default configuration for a box-box graph. Will modify configuration in-place and return it.

MCGraphs.Extractors.default_marker_genes_configuration — Function

default_marker_genes_configuration(
    daf::DafReader,
    [configuration::HeatmapGraphConfiguration = HeatmapGraphConfiguration();
    type_property::Maybe{AbstractString},
    min_significant_fold::Real = 0.5,
    max_significant_fold::Real = 3.0],
)::HeatmapGraphConfiguration

Return a default configuration for a markers heatmap graph. Will modify configuration in-place and return it.

Genes whose (absolute) fold factor (log base 2 of the ratio between the expression level and the median of the population) is less than min_significant_fold are colored in white. The color scale continues until max_significant_fold .

Index

MCGraphs.Extractors

MCGraphs.Extractors.default_box_box_configuration

MCGraphs.Extractors.default_gene_gene_configuration

MCGraphs.Extractors.default_marker_genes_configuration

MCGraphs.Extractors.extract_box_box_data

MCGraphs.Extractors.extract_categorical_color_palette

MCGraphs.Extractors.extract_gene_gene_data

MCGraphs.Extractors.extract_marker_genes_data