Plotters

MCGraphs.Plotters — Module

Generate complete graphs from a metacells Daf for standard graphs.

Graphs

MCGraphs.Plotters.plot_gene_gene — Function

plot_gene_gene(
    daf::DafReader
    [configuration::PointsGraphConfiguration = PointsGraphConfiguration()];
    x_gene::AbstractString,
    y_gene::AbstractString,
    [axis::QueryString = "metacell",
    color::Maybe{QueryString} = "type",
    min_significant_gene_UMIs::Integer = 40,
    gene_fraction_regularization::Maybe{AbstractFloat} = 1e-5,
    colors::Maybe{QueryString} = "/ type : color",
    entries_hovers::Maybe{FrameColumns} = ["total_UMIs" => "=", "type" => "="],
    genes_hovers::Maybe{FrameColumns} = nothing,]
)::PointsGraph

Generate a complete gene-gene graph using extract_gene_gene_data , extract_categorical_color_palette , and default_gene_gene_configuration .

By default, we look at gene expression per metacell; you can override this using the axis parameter. Each point in the graph is a gene which has a robust comparable expression (at least min_significant_gene_UMIs between both axis entries).

MCGraphs.Plotters.plot_box_box — Function

plot_box_box(
    daf::DafReader,
    [configuration::PointsGraphConfiguration = PointsGraphConfiguration()];
    x_box::AbstractString,
    y_box::AbstractString,
    [min_significant_gene_UMIs::Integer = 40,
    max_box_span::AbstractFloat = 2.0,
    gene_fraction_regularization::AbstractFloat = 1e-5,
    confidence::AbstractFloat = 0.9],
)::PointsGraph

Generate a complete box-box graph using extract_box_box_data and default_box_box_configuration .

Each point in the graph is a gene which has a robust comparable expression (at least min_significant_gene_UMIs between both spgeres). A line is attached to each point showing the confidence modification used when deciding on grouping of metacells into boxes.

MCGraphs.Plotters.plot_marker_genes — Function

function plot_marker_genes(
    daf::DafReader,
    [configuration::HeatmapGraphConfiguration = HeatmapGraphConfiguration();
    axis::abstractstring = "metacell",
    gene_names::maybe{vector{abstractstring}} = nothing,
    max_marker_genes::integer = 100,
    gene_fraction_regularization::abstractfloat = 1e-5,
    type_property::maybe{abstractstring} = "type",
    expression_annotations::maybe{querycolumns} = nothing,
    gene_annotations::maybe{querycolumns} = ["is_lateral", "divergence"],
    min_significant_fold::Real = 3.0,
    max_significant_fold::Real = 5.0,
    reorder_by_type::Bool = false]
)::HeatmapGraph

Generate a marker genes graph. This shows the genes that most distinguish between metacells (or profiles using another axis). Type annotations are added based on the type_property (which should name an axis with a color property), and optional expression_annotations and gene_annotations .

If gene_names is specified, these genes will always appear in the graph. This list is supplemented with additional is_marker genes to show at least min_marker_genes . A number of strongest such genes is chosen from each profile, such that the total number of these genes together with the forced gene_names is at least min_marker_genes .

The data is clustered to show the structure of both genes and metacells (or profiles using another axis). If reorder_by_type is specified, and type_property is not be nothing , then the profiles are reordered so that each type is contiguous.

Genes whose (absolute) fold factor (log base 2 of the ratio between the expression level and the median of the population) is less than min_significant_fold are colored in white. The color scale continues until max_significant_fold .

Index

MCGraphs.Plotters

MCGraphs.Plotters.plot_box_box

MCGraphs.Plotters.plot_gene_gene

MCGraphs.Plotters.plot_marker_genes