Renderers

MCGraphs.Renderers — Module

Render interactive or static graphs.

This provides a selection of basic graph types needed for metacells visualization. For each one, we define a struct containing all the data for the graph, and a separate struct containing the configuration of the graph. The rendering function takes both and either returns a JavaScript blob needed to create the interactive graph, or writes the graph to a file.

MCGraphs.Renderers.render — Function

render(
    data::SomeGraphData,
    configuration::SomeGraphConfiguration = SomeGraphConfiguration()
)::Nothing

Render a graph given its data and configuration. The implementation depends on the specific graph.

MCGraphs.Renderers.GraphConfiguration — Type

@kwdef mutable struct GraphConfiguration <: ObjectWithValidation
    file::Maybe{AbstractString} = nothing
    title::AbstractString = ""
    width::Maybe{Int} = nothing
    height::Maybe{Int} = nothing
end

Generic configuration that applies to any graph.

If output_file is specified, it is the path of a file to write the graph into (ending with .png or .svg ). If show_interactive is set, then generate an interactive graph (in a Jupyter notebook). One of output_file and show_interactive must be specified.

If specified, graph_title is used for the whole graph.

The optional graph_width and graph_height are in pixels, that is, 1/96 of an inch.

MCGraphs.Renderers.AxisConfiguration — Type

@kwdef mutable struct AxisConfiguration <: ObjectWithValidation
    title::AbstractString = ""
    minimum::Maybe{Real} = nothing
    maximum::Maybe{Real} = nothing
end

Generic configuration for a graph axis. Everything is optional; by default, the title is empty and the minimum and maximum are computed automatically from the data.

MCGraphs.Renderers.ValuesOrientation — Type

The orientation of the values axis in a distribution or bar graph:

HorizontalValues - The values are the X axis

VerticalValues - The values are the Y axis (the default).

Distribution

MCGraphs.Renderers.DistributionGraphConfiguration — Type

@kwdef mutable struct DistributionGraphConfiguration <: ObjectWithValidation
    graph::GraphConfiguration = GraphConfiguration()
    shape::DistributionShapeConfiguration = DistributionShapeConfiguration()
    orientation::ValuesOrientation = VerticalValues
    color::Maybe{AbstractString} = nothing
    values_axis::AxisConfiguration = AxisConfiguration()
    trace_title::Maybe{AbstractString} = nothing
end

Configure a graph for showing a distribution.

The optional color will be chosen automatically if not specified.

MCGraphs.Renderers.DistributionShapeConfiguration — Type

@kwdef mutable struct DistributionShapeConfiguration <: ObjectWithValidation
    show_box::Bool = true
    show_violin::Bool = false
    show_curve::Bool = false
    show_outliers::Bool = false
end

Configure the shape of a distribution graph.

If show_box , show a box graph.

If show_violin , show a violin graph.

If show_curve , show a density curve.

You can combine the above; however, a density curve is just the positive side of a violin graph, so you can't combine the two.

In addition to the (combination) of the above, if show_outliers , also show the extreme (outlier) points.

MCGraphs.Renderers.DistributionGraphData — Type

@kwdef mutable struct DistributionGraphData
    values::AbstractVector{<:Real}
    title::Maybe{AbstractString} = nothing
end

The data for a distribution graph, which is simply a vector of values.

Index

MCGraphs.Renderers

MCGraphs.Renderers.AxisConfiguration

MCGraphs.Renderers.DistributionGraphConfiguration

MCGraphs.Renderers.DistributionGraphData

MCGraphs.Renderers.DistributionShapeConfiguration

MCGraphs.Renderers.GraphConfiguration

MCGraphs.Renderers.ValuesOrientation

MCGraphs.Renderers.render