Validations

MCGraphs.Validations — Module

Validate user input.

Rendering graphs requires two objects: data and configuration. Both objects need to be internally consistent. This is especially relevant for the graph configuration. When creating UI for filling in these objects, we can in general easily validate each field on its own (e.g., ensure that a "color" field contains a valid color name). To ensure the overall object is consistent, we provide overall-type-specific validation functions that can be invoked by the UI to inform the user if the combination of (individually valid) field values is not valid for some reason.

MCGraphs.Validations.ObjectWithValidation — Type

A common type for objects that support validation, that is, that one can invoke validate_object on.

MCGraphs.Validations.validate_object — Function

validate_object(object::ObjectWithValidation)::Maybe{AbstractString}

Validate all field values of an object are compatible with each other, assuming each one is valid on its own. Returns nothing for a valid object and an error message if something is wrong. By default, this returns nothing .

This can be used by GUI widgets to validate the object as a whole (as opposed to validating each field based on its type).

MCGraphs.Validations.assert_valid_object — Function

assert_valid_object(object_with_validation::ObjectWithValidation)::Maybe{AbstractString}

This will @assert that the object_with_validation is valid (that is, validate_object will return nothing for it). This is used in the back-end (graph rendering) code. It is recommended that the front-end (UI) code will invoke validate_object and ensure the user fixes problems before invoking the back-end code.

Index

MCGraphs.Validations

MCGraphs.Validations.ObjectWithValidation

MCGraphs.Validations.assert_valid_object

MCGraphs.Validations.validate_object