MCGraphs

MCGraphs.MCGraphs — Module

Generate graphs for visualizing scRNA-seq metacells data in a Daf data set. The symbols from the main sub-modules are re-exported from the main MCGraphs namespace. These sub-modules are:

Validations - general API allowing controllers to validate use input.

Renderers - TODO: functions that actually render graphs given their data and configuration.

Extractors - TODO: functions that extract graph data from metacells Daf data sets.

Plotters - TODO: functions that combine extraction and rendering, creating a graph in one call.

Controllers - TODO: UI elements for specifying graph data or configuration.

Widgets - TODO: Combine multiple controllers for generating a complete graph.

Menus - TODO: A tree of menus for selecting and generating a single graph or a dashboard of multiple graphs.

Index

MCGraphs.MCGraphs

MCGraphs.Renderers

MCGraphs.Validations

MCGraphs.Renderers.AbstractGraphConfiguration

MCGraphs.Renderers.AbstractGraphData

MCGraphs.Renderers.AxisConfiguration

MCGraphs.Renderers.BandRegionConfiguration

MCGraphs.Renderers.BandsConfiguration

MCGraphs.Renderers.CdfDirection

MCGraphs.Renderers.CdfGraphConfiguration

MCGraphs.Renderers.CdfGraphData

MCGraphs.Renderers.CdfsGraphConfiguration

MCGraphs.Renderers.CdfsGraphData

MCGraphs.Renderers.DistributionGraphConfiguration

MCGraphs.Renderers.DistributionGraphData

MCGraphs.Renderers.DistributionStyleConfiguration

MCGraphs.Renderers.DistributionsGraphConfiguration

MCGraphs.Renderers.DistributionsGraphData

MCGraphs.Renderers.GraphConfiguration

MCGraphs.Renderers.LineGraphConfiguration

MCGraphs.Renderers.LineGraphData

MCGraphs.Renderers.LineStyleConfiguration

MCGraphs.Renderers.LinesGraphConfiguration

MCGraphs.Renderers.LinesGraphData

MCGraphs.Renderers.PointsGraphConfiguration

MCGraphs.Renderers.PointsGraphData

MCGraphs.Renderers.PointsStyleConfiguration

MCGraphs.Renderers.Stacking

MCGraphs.Renderers.ValuesOrientation

MCGraphs.Validations.ObjectWithValidation

MCGraphs.Renderers.render

MCGraphs.Validations.assert_valid_object

MCGraphs.Validations.validate_object