Extractors

MCGraphs.Extractors — Module

Extract data from a metacells Daf for standard graphs.

Data

MCGraphs.Extractors.extract_metacells_gene_gene_data — Function

extract_metacells_gene_gene_data(
    daf::DafReader;
    x_gene::AbstractString,
    y_gene::AbstractString,
    min_significant_gene_UMIs::Integer = 40,
    color_query::Maybe{QueryString} = ": type => color",
    colors_title::Maybe{AbstractString} = "Type",
    metacells_hovers::Maybe{FrameColumns} = ["type" => "="],
)::PointsGraphData

Extract the data for a metacells gene-gene graph from the daf data. The X coordinate of each point is the fraction of the x_gene and the Y coordinate of each gene is the fraction of the y_gene in each of the metacells.

We ignore genes that don't have at least min_significant_gene_UMIs between both entries.

If a colors_query is specified, it can be a suffix of a query that fetches a value for each metacell, or a full query that groups by metacells. By default we color by the type of the metacell. The colors_title is used for the legend.

For each metacell point, the hover will include the metacells_hovers per-metacell data.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

type (required): Distinct named biological cell states.

Vectors

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of all the genes in each metacell.

metacell @ type ::AbstractString (optional): The type each metacell belongs to.

type @ color ::AbstractString (optional): A unique color for each type for graphs.

Matrices

gene, metacell @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each metacell.

gene, metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

Additional inputs may be used depending to the query parameter(s).

MCGraphs.Extractors.extract_boxes_gene_gene_data — Function

extract_boxes_gene_gene_data(
    daf::DafReader;
    x_gene::AbstractString,
    y_gene::AbstractString,
    min_significant_gene_UMIs::Integer = 40,
    color_query::Maybe{QueryString} = ": type => color",
    colors_title::Maybe{AbstractString} = "Type",
    boxes_hovers::Maybe{FrameColumns} = ["type" => "="],
)::PointsGraphData

Extract the data for a boxes gene-gene graph from the daf data. The X coordinate of each point is the fraction of the x_gene and the Y coordinate of each gene is the fraction of the y_gene in each of the boxes.

We ignore genes that don't have at least min_significant_gene_UMIs between both entries.

If a colors_query is specified, it can be a suffix of a query that fetches a value for each box, or a full query that groups by boxes. By default we color by the type of the box. The colors_title is used for the legend.

For each box point, the hover will include the boxes_hovers per-box data.

Inputs

Axes

gene (required): Sequenced genes.

box (required): Distinct groups of metacells with "very close" estimated cell state.

type (required): Distinct named biological cell states.

Vectors

box @ neighborhood.main ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of all the genes in each box.

box @ type ::AbstractString (optional): The type each box belongs to.

type @ color ::AbstractString (optional): A unique color for each type for graphs.

Matrices

gene, box @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each box.

gene, box @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each box.

Additional inputs may be used depending to the query parameter(s).

MCGraphs.Extractors.extract_box_box_data — Function

function extract_box_box_data(
    daf::DafReader;
    x_box::AbstractString,
    y_box::AbstractString,
    [min_significant_gene_UMIs::Integer = 40,
    max_box_span::AbstractFloat = 2.0,
    gene_fraction_regularization::AbstractFloat = 1.0e-5],
    fold_confidence::AbstractFloat = 0.9,
)::PointsGraphData

Extract the data for a box-box graph. This shows why two boxes were not merged (or, if given the same box name twice, why the box was merged).

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

box (required): Distinct groups of metacells with "very close" estimated cell state.

neighborhood (required): Overlapping groups of boxes with "close" estimated cell states.

Vectors

metacell @ box ::AbstractString (required): The unique box each metacell belongs to.

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of all the genes in each metacell.

gene @ is_lateral ::Bool (required): A mask of genes that are lateral to the biological behaviors of interest.

gene @ divergence ::Union{Float32, Float64} (required): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

box @ neighborhood.main ::AbstractString (required): The unique main neighborhood of each box.

neighborhood @ span ::Union{Float32, Float64} (required): The span (fold factor) used to compute the neighborhood.

Matrices

gene, metacell @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each metacell.

gene, metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

gene, neighborhood @ is_correlated ::Bool (required): Which genes are correlated in each neighborhood.

box, neighborhood @ is_member ::Bool (required): A mask of the member boxes of each neighborhood.

MCGraphs.Extractors.extract_metacells_marker_genes_data — Function

function extract_metacells_marker_genes_data(
    daf::DafReader;
    [gene_names::Maybe{Vector{AbstractString}} = nothing,
    max_marker_genes::Integer = 100,
    gene_fraction_regularization::AbstractFloat = 1.0e-5,
    type_annotation::Bool = true,
    expression_annotations::Maybe{FrameColumns} = nothing,
    gene_annotations::Maybe{FrameColumns} = ["is_lateral", "divergence"],
    reorder_by_type::Bool = true]
)::HeatmapGraphData

Extract the data for a metacells marker genes graph. This shows the genes that most distinguish between metacells. If set, type_annotation is is added based on the type of each metacell. Optional expression_annotations and gene_annotations are added as well.

If gene_names is specified, these genes will always appear in the graph. This list is supplemented with additional is_marker genes to show at least min_marker_genes . A number of strongest such genes is chosen from each profile, such that the total number of these genes together with the forced gene_names is at least min_marker_genes .

If named_gene_annotation is set, and any gene_names were specified, then this is added as a gene annotation called "named".

The data is clustered to show the structure of both genes and metacells. If reorder_by_type is specified, then the profiles are reordered so that each type is contiguous.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

type (optional): Distinct named biological cell states.

Vectors

gene @ divergence ::Union{Float32, Float64} (required): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

gene @ is_lateral ::Bool (optional): A mask of genes that are lateral to the biological behaviors of interest.

box @ type ::AbstractString (optional): The type each box belongs to.

type @ color ::AbstractString (optional): A unique color for each type for graphs.

Matrices

gene, box @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each box.

Additional inputs may be used depending to the query parameter(s).

MCGraphs.Extractors.extract_boxes_marker_genes_data — Function

function extract_boxes_marker_genes_data(
    daf::DafReader;
    [gene_names::Maybe{Vector{AbstractString}} = nothing,
    max_marker_genes::Integer = 100,
    gene_fraction_regularization::AbstractFloat = 1.0e-5,
    type_annotation::Bool = true,
    expression_annotations::Maybe{FrameColumns} = nothing,
    named_gene_annotation::Bool = true,
    gene_annotations::Maybe{FrameColumns} = ["is_lateral", "divergence"],
    reorder_by_type::Bool = true]
)::HeatmapGraphData

Extract the data for a boxes marker genes graph. This shows the genes that most distinguish between boxes. If set, type_annotation is is added based on the type of each box. Optional expression_annotations and gene_annotations are added as well.

If gene_names is specified, these genes will always appear in the graph. This list is supplemented with additional is_marker genes to show at least min_marker_genes . A number of strongest such genes is chosen from each profile, such that the total number of these genes together with the forced gene_names is at least min_marker_genes .

If named_gene_annotation is set, and any gene_names were specified, then this is added as a gene annotation called "named".

The data is clustered to show the structure of both genes and boxes. If reorder_by_type is specified, and type_property is not be nothing , then the profiles are reordered so that each type is contiguous.

Inputs

Axes

gene (required): Sequenced genes.

box (required): Distinct groups of metacells with "very close" estimated cell state.

type (optional): Distinct named biological cell states.

Vectors

gene @ divergence ::Union{Float32, Float64} (required): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

gene @ is_lateral ::Bool (optional): A mask of genes that are lateral to the biological behaviors of interest.

box @ type ::AbstractString (optional): The type each box belongs to.

type @ color ::AbstractString (optional): A unique color for each type for graphs.

Matrices

gene, box @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each box.

Additional inputs may be used depending to the query parameter(s).

Configuration

MCGraphs.Extractors.default_gene_gene_configuration — Function

default_gene_gene_configuration(
    daf::DafReader,
    configuration = PointsGraphConfiguration();
    [color_query::Maybe{QueryString} = "/ type : color",
    gene_fraction_regularization::Maybe{AbstractFloat} = 1.0e-5]
)::PointsGraphConfiguration

Return a default configuration for a gene-gene graph. This applies the log scale to the axes, and sets up the color palette. Will modify configuration in-place and return it.

MCGraphs.Extractors.default_box_box_configuration — Function

default_box_box_configuration(
    [configuration::PointsGraphConfiguration = PointsGraphConfiguration()];
    x_box::AbstractString,
    y_box::AbstractString,
    x_neighborhood::AbstractString,
    [max_box_span::AbstractFloat = 2.0,
    gene_fraction_regularization::AbstractFloat = 1.0e-5],
)::PointsGraphConfiguration

Return a default configuration for a box-box graph. Will modify configuration in-place and return it.

MCGraphs.Extractors.default_marker_genes_configuration — Function

default_marker_genes_configuration(
    daf::DafReader,
    [configuration::HeatmapGraphConfiguration = HeatmapGraphConfiguration();
    type_annotation::Bool = true,
    min_significant_fold::Real = 0.5,
    max_significant_fold::Real = 3.0],
)::HeatmapGraphConfiguration

Return a default configuration for a markers heatmap graph. Will modify configuration in-place and return it.

Genes whose (absolute) fold factor (log base 2 of the ratio between the expression level and the median of the population) is less than min_significant_fold are colored in white. The color scale continues until max_significant_fold .

If type_annotation is set, sets up the color palette for the type annotations.

Inputs

Axes

type (optional): Distinct named biological cell states.

Vectors

type @ color ::AbstractString (optional): A unique color for each type for graphs.

Index

MCGraphs.Extractors

MCGraphs.Extractors.default_box_box_configuration

MCGraphs.Extractors.default_gene_gene_configuration

MCGraphs.Extractors.default_marker_genes_configuration

MCGraphs.Extractors.extract_box_box_data

MCGraphs.Extractors.extract_boxes_gene_gene_data

MCGraphs.Extractors.extract_boxes_marker_genes_data

MCGraphs.Extractors.extract_metacells_gene_gene_data

MCGraphs.Extractors.extract_metacells_marker_genes_data