Plotters

MCGraphs.Plotters — Module

Generate complete graphs from a metacells Daf for standard graphs.

Graphs

MCGraphs.Plotters.plot_gene_gene — Function

plot_gene_gene(
    daf::DafReader;
    x_gene::AbstractString,
    y_gene::AbstractString,
    axis::QueryString = "metacell",
    color::Maybe{QueryString} = "type",
    min_significant_gene_UMIs::Integer = 40,
    gene_fraction_regularization::Maybe{AbstractFloat} = 1e-5,
    colors::Maybe{QueryString} = "/ type : color",
    entries_hovers::Maybe{QueryColumns} = ["total_UMIs" => "=", "type" => "="],
    genes_hovers::Maybe{QueryColumns} = nothing,
)::Graph

Generate a complete gene-gene graph using extract_gene_gene_data , extract_categorical_color_palette , and default_gene_gene_configuration .

By default, we look at gene expression per metacell; you can override this using the axis parameter. Each point in the graph is a gene which has a robust comparable expression (at least min_significant_gene_UMIs between both axis entries).

MCGraphs.Plotters.plot_sphere_sphere — Function

plot_sphere_sphere(
    daf::DafReader;
    x_sphere::AbstractString,
    y_sphere::AbstractString,
    min_significant_gene_UMIs::Integer = 40,
    max_sphere_diameter::AbstractFloat = 2.0,
    gene_fraction_regularization::AbstractFloat = 1e-5,
    confidence::AbstractFloat = 0.9,
)::Graph

Generate a complete sphere-sphere graph using extract_sphere_sphere_data and default_sphere_sphere_configuration .

Each point in the graph is a gene which has a robust comparable expression (at least min_significant_gene_UMIs between both spgeres). A line is attached to each point showing the confidence modification used when deciding on grouping of metacells into spheres.

Index

MCGraphs.Plotters

MCGraphs.Plotters.plot_gene_gene

MCGraphs.Plotters.plot_sphere_sphere