Plotters

MCGraphs.Plotters — Module

Generate complete graphs from a metacells Daf for standard graphs.

Graphs

MCGraphs.Plotters.plot_metacells_gene_gene — Function

plot_metacells_gene_gene(
    daf::DafReader
    [configuration::PointsGraphConfiguration = PointsGraphConfiguration()];
    x_gene::AbstractString,
    y_gene::AbstractString,
    [min_significant_gene_UMIs::Integer = 40,
    gene_fraction_regularization::Maybe{AbstractFloat} = 1.0e-5,
    color_query::Maybe{QueryString} = ": type => color",
    metacells_hovers::Maybe{FrameColumns} = ["type" => "="]]
)::PointsGraph

Generate a complete metacells gene-gene graph using extract_metacells_gene_gene_data and default_gene_gene_configuration .

Each point in the graph is a gene which has a robust comparable expression (at least min_significant_gene_UMIs between both metacells).

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

type (required): Distinct named biological cell states.

Vectors

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of all the genes in each metacell.

metacell @ type ::AbstractString (optional): The type each metacell belongs to.

type @ color ::AbstractString (optional): A unique color for each type for graphs.

Matrices

gene, metacell @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each metacell.

gene, metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

Additional inputs may be used depending to the query parameter(s).

MCGraphs.Plotters.plot_boxes_gene_gene — Function

plot_boxes_gene_gene(
    daf::DafReader
    [configuration::PointsGraphConfiguration = PointsGraphConfiguration()];
    x_gene::AbstractString,
    y_gene::AbstractString,
    [min_significant_gene_UMIs::Integer = 40,
    gene_fraction_regularization::Maybe{AbstractFloat} = 1.0e-5,
    color_query::Maybe{QueryString} = ": type => color",
    boxes_hovers::Maybe{FrameColumns} = ["type" => "="]]
)::PointsGraph

Generate a complete boxes gene-gene graph using extract_boxes_gene_gene_data and default_gene_gene_configuration .

Each point in the graph is a gene which has a robust comparable expression (at least min_significant_gene_UMIs between both boxes).

Inputs

Axes

gene (required): Sequenced genes.

box (required): Distinct groups of metacells with "very close" estimated cell state.

type (required): Distinct named biological cell states.

Vectors

box @ neighborhood.main ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of all the genes in each box.

box @ type ::AbstractString (optional): The type each box belongs to.

type @ color ::AbstractString (optional): A unique color for each type for graphs.

Matrices

gene, box @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each box.

gene, box @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each box.

Additional inputs may be used depending to the query parameter(s).

MCGraphs.Plotters.plot_box_box — Function

plot_box_box(
    daf::DafReader,
    [configuration::PointsGraphConfiguration = PointsGraphConfiguration()];
    x_box::AbstractString,
    y_box::AbstractString,
    [min_significant_gene_UMIs::Integer = 40,
    max_box_span::AbstractFloat = 2.0,
    gene_fraction_regularization::AbstractFloat = 1.0e-5],
)::PointsGraph

Generate a complete box-box graph using extract_box_box_data and default_box_box_configuration .

Each point in the graph is a gene which has a robust comparable expression (at least min_significant_gene_UMIs between both spgeres). A line is attached to each point showing the confidence modification used when deciding on grouping of metacells into boxes.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

box (required): Distinct groups of metacells with "very close" estimated cell state.

neighborhood (required): Overlapping groups of boxes with "close" estimated cell states.

Vectors

metacell @ box ::AbstractString (required): The unique box each metacell belongs to.

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of all the genes in each metacell.

gene @ is_lateral ::Bool (required): A mask of genes that are lateral to the biological behaviors of interest.

gene @ divergence ::Union{Float32, Float64} (required): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

box @ neighborhood.main ::AbstractString (required): The unique main neighborhood of each box.

neighborhood @ span ::Union{Float32, Float64} (required): The span (fold factor) used to compute the neighborhood.

Matrices

gene, metacell @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each metacell.

gene, metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

gene, neighborhood @ is_correlated ::Bool (required): Which genes are correlated in each neighborhood.

box, neighborhood @ is_member ::Bool (required): A mask of the member boxes of each neighborhood.

MCGraphs.Plotters.plot_metacells_marker_genes — Function

function plot_metacells_marker_genes(
    daf::DafReader,
    [configuration::HeatmapGraphConfiguration = HeatmapGraphConfiguration();
    gene_names::Maybe{Vector{AbstractString}} = nothing,
    max_marker_genes::Integer = 100,
    gene_fraction_regularization::AbstractFloat = 1.0e-5,
    type_annotation::Bool = true,
    expression_annotations::Maybe{FrameColumns} = nothing,
    named_gene_annotation::Bool = true,
    gene_annotations::Maybe{FrameColumns} = ["is_lateral", "divergence"],
    reorder_by_type::Bool = true,
    min_significant_fold::Real = 0.5,
    max_significant_fold::Real = 3.0]
)::HeatmapGraph

Generate a metacells marker genes graph. This shows the genes that most distinguish between metacells (or profiles using another axis). Type annotations are added based on the type_property (which should name an axis with a color property), and optional expression_annotations and gene_annotations .

If gene_names is specified, these genes will always appear in the graph. This list is supplemented with additional is_marker genes to show at least min_marker_genes . A number of strongest such genes is chosen from each profile, such that the total number of these genes together with the forced gene_names is at least min_marker_genes .

The data is clustered to show the structure of both genes and metacells (or profiles using another axis). If reorder_by_type is specified, and type_property is not be nothing , then the profiles are reordered so that each type is contiguous.

Genes whose (absolute) fold factor (log base 2 of the ratio between the expression level and the median of the population) is less than min_significant_fold are colored in white. The color scale continues until max_significant_fold .

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

gene (required): Sequenced genes.

type (optional): Distinct named biological cell states.

Vectors

box @ type ::AbstractString (optional): The type each box belongs to.

gene @ is_lateral ::Bool (optional): A mask of genes that are lateral to the biological behaviors of interest.

gene @ divergence ::Union{Float32, Float64} (required): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

type @ color ::AbstractString (optional): A unique color for each type for graphs.

Matrices

gene, box @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each box.

Additional inputs may be used depending to the query parameter(s).

MCGraphs.Plotters.plot_boxes_marker_genes — Function

function plot_boxes_marker_genes(
    daf::DafReader,
    [configuration::HeatmapGraphConfiguration = HeatmapGraphConfiguration();
    gene_names::Maybe{Vector{AbstractString}} = nothing,
    max_marker_genes::Integer = 100,
    gene_fraction_regularization::AbstractFloat = 1.0e-5,
    type_annotation::Bool = true,
    expression_annotations::Maybe{FrameColumns} = nothing,
    named_gene_annotation::Bool = true,
    gene_annotations::Maybe{FrameColumns} = ["is_lateral", "divergence"],
    reorder_by_type::Bool = truem
    min_significant_fold::Real = 0.5,
    max_significant_fold::Real = 3.0]
)::HeatmapGraph

Generate a boxes marker genes graph. This shows the genes that most distinguish between boxes (or profiles using another axis). Type annotations are added based on the type_property (which should name an axis with a color property), and optional expression_annotations and gene_annotations .

If gene_names is specified, these genes will always appear in the graph. This list is supplemented with additional is_marker genes to show at least min_marker_genes . A number of strongest such genes is chosen from each profile, such that the total number of these genes together with the forced gene_names is at least min_marker_genes .

The data is clustered to show the structure of both genes and boxes (or profiles using another axis). If reorder_by_type is specified, and type_property is not be nothing , then the profiles are reordered so that each type is contiguous.

Genes whose (absolute) fold factor (log base 2 of the ratio between the expression level and the median of the population) is less than min_significant_fold are colored in white. The color scale continues until max_significant_fold .

Inputs

Axes

gene (required): Sequenced genes.

box (required): Distinct groups of metacells with "very close" estimated cell state.

type (optional): Distinct named biological cell states.

Vectors

box @ type ::AbstractString (optional): The type each box belongs to.

gene @ is_lateral ::Bool (optional): A mask of genes that are lateral to the biological behaviors of interest.

gene @ divergence ::Union{Float32, Float64} (required): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

type @ color ::AbstractString (optional): A unique color for each type for graphs.

Matrices

gene, box @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each box.

Additional inputs may be used depending to the query parameter(s).

Index

MCGraphs.Plotters

MCGraphs.Plotters.plot_box_box

MCGraphs.Plotters.plot_boxes_gene_gene

MCGraphs.Plotters.plot_boxes_marker_genes

MCGraphs.Plotters.plot_metacells_gene_gene

MCGraphs.Plotters.plot_metacells_marker_genes