Renderers

MCGraphs.Renderers — Module

Render interactive or static graphs.

This provides a selection of basic graph types needed for metacells visualization. For each one, we define a struct containing all the data for the graph, and a separate struct containing the configuration of the graph. The rendering function takes both and either returns a JavaScript blob needed to create the interactive graph, or writes the graph to a file.

MCGraphs.Renderers.render — Function

render(
    data::SomeGraphData,
    configuration::SomeGraphConfiguration = SomeGraphConfiguration()
)::Nothing

Render a graph given its data and configuration. The implementation depends on the specific graph.

Distribution

MCGraphs.Renderers.DistributionGraphData — Type

@kwdef struct DistributionGraphData
    values::AbstractVector{<:AbstractFloat}
    colors::Maybe{AbstractStringVector} = nothing
end

The data for a distribution graph. The values vector is mandatory. The colors vector is optional and not normally specified. It only has an effect if the DistributionGraphConfiguration specifies showing the points.

MCGraphs.Renderers.DistributionGraphConfiguration — Type

@kwdef struct DistributionGraphConfiguration <: ObjectWithValidation
    output::AbstractString = nothing
    title::AbstractString = nothing
    values_title::AbstractString = nothing
    horizontal::Bool = true
    violin::Bool = true
    box::Bool = false
    points::Bool = false
end

Configure a graph for showing a distribution.

output can be the path of a file (ending with .png or .svg ). If it is nothing (the default), then render a JavaScript blob for an interactive graph.

If specified, title is used for the whole graph and values_title is used for the values axis.

If horizontal (the default), the value would be the X axis and the density would be the Y axis; otherwise, this will be reversed.

If violin , show a violin plot. If curve , show a density curve (which is just the positive side of the violin plot, so you can't specify both curve and violin ). If box , show a box plot. If points , show the data points. Any other combination of these can be used as long as at least one of them is set. By defaults, shows just the curve plot.

The optional line_color , fill_color or point_color will be automatically chosen if not specified. The point_color is only used if showing the points . Also, if the graph data contains a color per point, then it overrides the configuration point_color .

Index

MCGraphs.Renderers

MCGraphs.Renderers.DistributionGraphConfiguration

MCGraphs.Renderers.DistributionGraphData

MCGraphs.Renderers.render