Renderers

MCGraphs.Renderers — Module

Render interactive or static graphs.

This provides a selection of basic graph types needed for metacells visualization. For each one, we define a struct containing all the data for the graph, and a separate struct containing the configuration of the graph. The rendering function takes both and either returns a JavaScript blob needed to create the interactive graph, or writes the graph to a file.

MCGraphs.Renderers.render — Function

render(
    data::SomeGraphData,
    configuration::SomeGraphConfiguration = SomeGraphConfiguration()
)::Nothing

Render a graph given its data and configuration. The implementation depends on the specific graph. The supported combinations of data and configuration are:

Data 	Configuration 	Description

DistributionGraphData 	 DistributionGraphConfiguration 	Graph of a single distribution.

DistributionsGraphData 	 DistributionsGraphConfiguration 	Graph of multiple distributions.

LineGraphData 	 LineGraphConfiguration 	Graph of a single line (function).

PointsGraphData 	 PointsGraphConfiguration 	Graph of points, possibly with edges between them.

MCGraphs.Renderers.GraphConfiguration — Type

@kwdef mutable struct GraphConfiguration <: ObjectWithValidation
    output_file::Maybe{AbstractString} = nothing
    show_interactive::Bool = false
    width::Maybe{Int} = nothing
    height::Maybe{Int} = nothing
    template::AbstractString = "simple_white"
    show_grid::Bool = true
    show_ticks::Bool = true
end

Generic configuration that applies to any graph.

If output_file is specified, it is the path of a file to write the graph into (ending with .png or .svg ). If show_interactive is set, then generate an interactive graph (in a Jupyter notebook). One of output_file and show_interactive must be specified.

The optional width and height are in pixels, that is, 1/96 of an inch.

By default, show_grid and show_ticks are set.

The default template is "simple white" which is the cleanest. The `show grid and show_ticks` can be used to disable the grid and/or ticks for an even cleaner (but less informative) look.

MCGraphs.Renderers.AxisConfiguration — Type

@kwdef mutable struct AxisConfiguration <: ObjectWithValidation
    minimum::Maybe{Real} = nothing
    maximum::Maybe{Real} = nothing
    log_scale::Bool,
end

Generic configuration for a graph axis. Everything is optional; by default, the minimum and maximum are computed automatically from the data.

If log_scale is set, the data must contain only positive values, and the axis is shown in log (base 10) scale. To help with finer-grained ratios, each 10x step is broken to three ~2.15 steps (which is "close enough" to 2x for intuitive reading of the ratios).

MCGraphs.Renderers.ValuesOrientation — Type

The orientation of the values axis in a distribution or a bars graph:

HorizontalValues - The values are the X axis

VerticalValues - The values are the Y axis (the default).

Distribution

MCGraphs.Renderers.DistributionGraphData — Type

@kwdef mutable struct DistributionGraphData
    graph_title::Maybe{AbstractString} = nothing
    value_axis_title::Maybe{AbstractString} = nothing
    trace_axis_title::Maybe{AbstractString} = nothing
    values::AbstractVector{<:Real}
    name::Maybe{AbstractString} = nothing
end

By default, all the titles are empty. You can specify the overall graph_title as well as the value_axis_title and the trace_axis_title . The optional name is used as the tick value for the distribution.

MCGraphs.Renderers.DistributionsGraphData — Type

@kwdef mutable struct DistributionsGraphData <: ObjectWithValidation
    graph_title::Maybe{AbstractString} = nothing
    value_axis_title::Maybe{AbstractString} = nothing
    trace_axis_title::Maybe{AbstractString} = nothing
    values::AbstractVector{<:AbstractVector{<:Real}}
    names::Maybe{AbstractStringVector} = nothing
    colors::Maybe{AbstractStringVector} = nothing
end

The data for a multiple distributions graph. By default, all the titles are empty. You can specify the overall graph_title as well as the value_axis_title and the trace_axis_title . If specified, the names and/or the colors vectors must contain the same number of elements as the number of vectors in the values .

MCGraphs.Renderers.DistributionGraphConfiguration — Type

@kwdef mutable struct DistributionGraphConfiguration <: ObjectWithValidation
    graph::GraphConfiguration = GraphConfiguration()
    style::DistributionStyleConfiguration = DistributionStyleConfiguration()
    value_axis::AxisConfiguration = AxisConfiguration()
end

Configure a graph for showing a distribution (with DistributionGraphData ) or several distributions (with DistributionsGraphData ).

The optional color will be chosen automatically if not specified. When showing multiple distributions, it is also possible to specify the color of each one in the DistributionsGraphData .

MCGraphs.Renderers.DistributionsGraphConfiguration — Type

@kwdef mutable struct DistributionsGraphConfiguration <: ObjectWithValidation
    graph::GraphConfiguration = GraphConfiguration()
    style::DistributionStyleConfiguration = DistributionStyleConfiguration()
    value_axis::AxisConfiguration = AxisConfiguration()
    show_legend::Bool = false
    overlay::Bool = false
end

Configure a graph for showing several distributions several distributions.

This is identical to DistributionGraphConfiguration with the addition of show_legend to show a legend. This is not set by default as it makes little sense unless overlay is also set. TODO: Implement overlay .

MCGraphs.Renderers.DistributionStyleConfiguration — Type

@kwdef mutable struct DistributionStyleConfiguration <: ObjectWithValidation
    show_box::Bool = true
    show_violin::Bool = false
    show_curve::Bool = false
    show_outliers::Bool = false
    color::Maybe{AbstractString} = nothing
end

Configure the style of a distribution graph.

If show_box , show a box graph.

If show_violin , show a violin graph.

If show_curve , show a density curve.

You can combine the above; however, a density curve is just the positive side of a violin graph, so you can't combine the two.

In addition to the (combination) of the above, if show_outliers , also show the extreme (outlier) points.

The color is chosen automatically by default. When showing multiple distributions, you can override it per each one in the DistributionsGraphData .

Lines

MCGraphs.Renderers.LineGraphData — Type

@kwdef mutable struct LineGraphData <: ObjectWithValidation
    graph_title::Maybe{AbstractString} = nothing
    x_axis_title::Maybe{AbstractString} = nothing
    y_axis_title::Maybe{AbstractString} = nothing
    xs::AbstractVector{<:Real}
    ys::AbstractVector{<:Real}
end

The data for a line graph (e.g. a CDF graph).

By default, all the titles are empty. You can specify the overall graph_title as well as the x_axis_title and y_axis_title for the axes.

The xs and ys vectors must be of the same size. A line will be drawn through all the points, and the area under the line may be filled.

MCGraphs.Renderers.LineGraphConfiguration — Type

@kwdef mutable struct LineGraphConfiguration <: ObjectWithValidation
    graph::GraphConfiguration = GraphConfiguration()
    x_axis::AxisConfiguration = AxisConfiguration()
    y_axis::AxisConfiguration = AxisConfiguration()
    style::LineStyleConfiguration = LineStyleConfiguration()
    vertical_bands::BandsConfiguration = BandsConfiguration()
    horizontal_bands::BandsConfiguration = BandsConfiguration()
end

Configure a graph for showing line plots.

MCGraphs.Renderers.LineStyleConfiguration — Type

@kwdef mutable struct LineStyleConfiguration <: ObjectWithValidation
    line_width::Maybe{Real} = 1.0
    fill_below::Bool = false
    line_is_dashed::Bool = false
    line_color::Maybe{AbstractString} = nothing
end

Configure a line in a graph.

By default, a solid line is shown; if line_is_dashed , the line will be dashed. If fill_below is set, the area below the line is filled. If the line_width is set to nothing , no line is shown (and fill_below must be set). By default, the line_color is chosen automatically.

Points

MCGraphs.Renderers.PointsGraphData — Type

@kwdef mutable struct PointsGraphData <: ObjectWithValidation
    graph_title::Maybe{AbstractString} = nothing
    x_axis_title::Maybe{AbstractString} = nothing
    y_axis_title::Maybe{AbstractString} = nothing
    xs::AbstractVector{<:Real}
    ys::AbstractVector{<:Real}
    colors::Maybe{Union{AbstractStringVector, AbstractVector{<:Real}}} = nothing
    sizes::Maybe{AbstractVector{<:Real}} = nothing
    hovers::Maybe{AbstractStringVector} = nothing
    border_colors::Maybe{Union{AbstractStringVector, AbstractVector{<:Real}}} = nothing
    border_sizes::Maybe{AbstractVector{<:Real}} = nothing
    edges::Maybe{AbstractVector{Tuple{<:Integer, <:Integer}}} = nothing
    edges_colors::Maybe{AbstractStringVector} = nothing
    edges_sizes::Maybe{<:AbstractVector{<:Real}} = nothing
end

The data for a scatter graph of points.

By default, all the titles are empty. You can specify the overall graph_title as well as the x_axis_title and y_axis_title for the axes.

The xs and ys vectors must be of the same size. If specified, the colors sizes and/or hovers vectors must also be of the same size. The colors can be either color names or a numeric value; if the latter, then the configuration's color_scale is used. Sizes are the diameter in pixels (1/96th of an inch). Hovers are only shown in interactive graphs (or when saving an HTML file).

The border_colors and border_sizes can be used to display additional data per point. The border size is in addition to the point size.

It is possible to draw straight edges between specific point pairs. In this case the edges_style of the PointsGraphConfiguration will be used, and the edges_colors and edges_sizes will override it per edge. The edges_colors are restricted to explicit colors, not a color scale.

MCGraphs.Renderers.PointsGraphConfiguration — Type

@kwdef mutable struct PointsGraphConfiguration <: ObjectWithValidation
    graph::GraphConfiguration = GraphConfiguration()
    x_axis::AxisConfiguration = AxisConfiguration()
    y_axis::AxisConfiguration = AxisConfiguration()
    style::PointsStyleConfiguration = PointsStyleConfiguration()
    border_style::PointsStyleConfiguration = PointsStyleConfiguration()
    edges_style::PointsConfiguration = PointsStyleConfiguration()
    vertical_bands::BandsConfiguration = BandsConfiguration()
    horizontal_bands::BandsConfiguration = BandsConfiguration()
    diagonal_bands::BandsConfiguration = BandsConfiguration()
end

Configure a graph for showing a scatter graph of points.

Using the vertical_bands , horizontal_bands and/or diagonal_bands you can partition the graph into regions. The diagonal_bands can only be used if both axes are linear or both axes are in log_scale ; they also unify the ranges of the X and Y axes. If the axes are in log_scale , the line_offset of the diagonal_bands are multiplicative instead of additive, and must be positive.

The border_style is used if the PointsGraphData contains either the border_colors and/or border_sizes . This allows displaying some additional data per point.

Note

You can't set the show_legend of the GraphConfiguration of a points graph. Instead you probably want to set the show_scale of the style (and/or of the border_style and/or edges_style ). In addition, the color scale options of the edges_style must not be set as edges_colors of PointsGraphData is restricted to explicit colors.

MCGraphs.Renderers.PointsStyleConfiguration — Type

@kwdef mutable struct PointsStyleConfiguration <: ObjectWithValidation
    size::Maybe{Real} = nothing
    color::Maybe{AbstractString} = nothing
    color_scale::Maybe{Union{
        AbstractString,
        AbstractVector{<:Tuple{<:Real, <:AbstractString}},
        AbstractVector{<:Tuple{<:AbstractString, <:AbstractString}}
    }} = nothing
    reverse_scale::Bool = false
    show_scale::Bool = false
end

Configure points in a graph. By default, the point size and color is chosen automatically (when this is applied to edges, the size is the width of the line). You can also override this by specifying sizes and colors in the PointsGraphData . If the data contains numeric color values, then the color_scale will be used instead; you can set reverse_scale to reverse it. You need to explicitly set show_scale to show its legend.

The color_scale can be the name of a standard one, a vector of (value, color) tuples for a continuous scale. If the values are numbers, the scale is continuous; if they are strings, this is a categorical scale.

Bands in graphs

MCGraphs.Renderers.BandsConfiguration — Type

@kwdef mutable struct BandsConfiguration <: ObjectWithValidation
    low::BandRegionConfiguration = BandRegionConfiguration(line_is_dashed = true)
    middle::BandRegionConfiguration = BandRegionConfiguration()
    high::BandRegionConfiguration = BandRegionConfiguration(line_is_dashed = true)
end

Configure the partition of the graph up to three band regions. The low and high are the "outer" regions (so their lines are at their border, dashed by default) and the middle is the "inner" region between them (so its line is inside it, solid by default).

MCGraphs.Renderers.BandRegionConfiguration — Type

@kwdef mutable struct BandRegionConfiguration <: ObjectWithValidation
    line_color::Maybe{AbstractString} = "black"
    line_width::Real = 1.0
    line_is_dashed::Bool = false
    fill_color::Maybe{AbstractString} = nothing
end

Configure a region of the graph defined by some band of values. The region only exists if line_offset is set. To actually show the region, either the line_color and/or the fill_color must be set; by default, just the line is shown. The line_width is in pixels (1/96th of an inch).

Index

MCGraphs.Renderers

MCGraphs.Renderers.AxisConfiguration

MCGraphs.Renderers.BandRegionConfiguration

MCGraphs.Renderers.BandsConfiguration

MCGraphs.Renderers.DistributionGraphConfiguration

MCGraphs.Renderers.DistributionGraphData

MCGraphs.Renderers.DistributionStyleConfiguration

MCGraphs.Renderers.DistributionsGraphConfiguration

MCGraphs.Renderers.DistributionsGraphData

MCGraphs.Renderers.GraphConfiguration

MCGraphs.Renderers.LineGraphConfiguration

MCGraphs.Renderers.LineGraphData

MCGraphs.Renderers.LineStyleConfiguration

MCGraphs.Renderers.PointsGraphConfiguration

MCGraphs.Renderers.PointsGraphData

MCGraphs.Renderers.PointsStyleConfiguration

MCGraphs.Renderers.ValuesOrientation

MCGraphs.Renderers.render