Renderers

MCGraphs.Renderers — Module

Render interactive or static graphs.

This provides a selection of basic graph types needed for metacells visualization. For each one, we define a struct containing all the data for the graph, and a separate struct containing the configuration of the graph. The rendering function takes both and either returns a JavaScript blob needed to create the interactive graph, or writes the graph to a file.

TODO:

Sizes legends.

Heatmaps.

Heatmap subplots as annotations for bar graph.

Heatmap subplots as annotations for heatmaps.

MCGraphs.Renderers.Graph — Type

The type of a figure we can display. This is a combination of some AbstractGraphData and AbstractGraphConfiguration , which we can pass to render to obtain a PlotlyFigure which Julia knows how to display.

The valid combinations of concrete data and configuration which we can render are:

AbstractGraphData 	 AbstractGraphConfiguration 	Description

PointsGraphData 	 PointsGraphConfiguration 	Graph of points, possibly with edges between them.

GridGraphData 	 GridGraphConfiguration 	Graph of a grid of points (e.g. for correlations).

LineGraphData 	 LineGraphConfiguration 	Graph of a single line (e.g. a function y=f(x)).

LinesGraphData 	 LinesGraphConfiguration 	Graph of multiple functions, possibly stacked.

CdfGraphData 	 CdfGraphConfiguration 	Graph of a single cumulative distribution function.

CdfsGraphData 	 CdfsGraphConfiguration 	Graph of multiple cumulative distribution functions.

DistributionGraphData 	 DistributionGraphConfiguration 	Graph of a single distribution.

DistributionsGraphData 	 DistributionsGraphConfiguration 	Graph of multiple distributions.

BarGraphData 	 BarGraphConfiguration 	Graph of a single set of bars (histogram).

BarsGraphData 	 BarsGraphConfiguration 	Graph of multiple sets of bars (histograms).

MCGraphs.Renderers.AbstractGraphData — Type

Common abstract base for all complete graph data types. See Graph .

MCGraphs.Renderers.AbstractGraphConfiguration — Type

Common abstract base for all complete graph configuration types. See Graph .

MCGraphs.Renderers.render — Function

render(
    data::AbstractGraphData,
    configuration::AbstractGraphConfiguration = ...,
)::PlotlyFigure

Render a graph given its data and configuration. The implementation depends on the specific graph. For each AbstractGraphData there is a matching AbstractGraphConfiguration (a default one is provided for the configuration ). The supported type pairs are:

AbstractGraphData 	 AbstractGraphConfiguration 	Description

PointsGraphData 	 PointsGraphConfiguration 	Graph of points, possibly with edges between them.

GridGraphData 	 GridGraphConfiguration 	Graph of a grid of points (e.g. for correlations).

LineGraphData 	 LineGraphConfiguration 	Graph of a single line (e.g. a function y=f(x)).

LinesGraphData 	 LinesGraphConfiguration 	Graph of multiple functions, possibly stacked.

CdfGraphData 	 CdfGraphConfiguration 	Graph of a single cumulative distribution function.

CdfsGraphData 	 CdfsGraphConfiguration 	Graph of multiple cumulative distribution functions.

DistributionGraphData 	 DistributionGraphConfiguration 	Graph of a single distribution.

DistributionsGraphData 	 DistributionsGraphConfiguration 	Graph of multiple distributions.

BarGraphData 	 BarGraphConfiguration 	Graph of a single set of bars (histogram).

BarsGraphData 	 BarsGraphConfiguration 	Graph of multiple sets of bars (histograms).

This returns a PlotlyFigure which can be displayed directly, or converted to JSON for transfer to other programming languages (Python or R)

MCGraphs.Renderers.PlotlyFigure — Type

The type of a rendered graph which Julia knows how to display. See render .

A plotly figure contains everything needed to display an interactive graph (or generate a static one on disk). It can also be converted to a JSON string for handing it over to a different programming language (e.g., to be used to display the interactive graph in a Python Jupyter notebook, given an appropriate wrapper code).

Points

MCGraphs.Renderers.PointsGraphData — Type

@kwdef mutable struct PointsGraphData <: AbstractGraphData
    graph_title::Maybe{AbstractString} = nothing
    x_axis_title::Maybe{AbstractString} = nothing
    y_axis_title::Maybe{AbstractString} = nothing
    points_colors_title::Maybe{AbstractString} = nothing
    points_sizes_title::Maybe{AbstractString} = nothing
    border_colors_title::Maybe{AbstractString} = nothing
    border_sizes_title::Maybe{AbstractString} = nothing
    points_xs::AbstractVector{<:Real}
    points_ys::AbstractVector{<:Real}
    points_colors::Maybe{Union{AbstractVector{<:AbstractString}, AbstractVector{<:Real}}} = nothing
    sizes::Maybe{AbstractVector{<:Real}} = nothing
    hovers::Maybe{AbstractVector{<:AbstractString}} = nothing
    borders_colors::Maybe{Union{AbstractVector{<:AbstractString}, AbstractVector{<:Real}}} = nothing
    borders_sizes::Maybe{AbstractVector{<:Real}} = nothing
    edges::Maybe{AbstractVector{Tuple{<:Integer, <:Integer}}} = nothing
    edges_colors::Maybe{AbstractVector{<:AbstractString}} = nothing
    edges_sizes::Maybe{AbstractVector{<:Real}} = nothing
end

The data for a scatter graph of points.

By default, all the titles are empty. You can specify the overall graph_title as well as the x_axis_title and y_axis_title for the axes.

The points_xs and points_ys vectors must be of the same size. If specified, the points_colors sizes and/or hovers vectors must also be of the same size. The points_colors can be either color names or a numeric value; if the latter, then the configuration's color_palette is used. Sizes are the diameter in pixels (1/96th of an inch). Hovers are only shown in interactive graphs (or when saving an HTML file).

The borders_colors and borders_sizes can be used to display additional data per point. The border size is in addition to the point size.

The points_colors_title , points_sizes_title , border_colors_title and border_sizes_title are only used if show_scale is set for the relevant color scales. You can't specify show_scale if there is no points_colors data or if the points_colors contain explicit color names.

It is possible to draw straight edges between specific point pairs. In this case the edges of the PointsGraphConfiguration will be used, and the edges_colors and edges_sizes will override it per edge. The edges_colors are restricted to explicit colors, not a color scale.

A point (or a point border, or an edge) with a zero size and/or an empty string color (either from the data or from a categorical color_palette ) will not be shown.

MCGraphs.Renderers.PointsGraphConfiguration — Type

@kwdef mutable struct PointsGraphConfiguration <: AbstractGraphConfiguration
    graph::GraphConfiguration = GraphConfiguration()
    x_axis::AxisConfiguration = AxisConfiguration()
    y_axis::AxisConfiguration = AxisConfiguration()
    points::PointsConfiguration = PointsConfiguration()
    border::PointsConfiguration = PointsConfiguration()
    edges::PointsConfiguration = PointsConfiguration()
    vertical_bands::BandsConfiguration = BandsConfiguration()
    horizontal_bands::BandsConfiguration = BandsConfiguration()
    diagonal_bands::BandsConfiguration = BandsConfiguration()
end

Configure a graph for showing a scatter graph of points.

Using the vertical_bands , horizontal_bands and/or diagonal_bands you can partition the graph into regions. The diagonal_bands can only be used if both axes are linear or both axes are in log scale; they also unify the ranges of the X and Y axes. If the axes are in log scale, the offset of the diagonal_bands are multiplicative instead of additive, and must be positive.

The border is used if the PointsGraphData contains either the borders_colors and/or borders_sizes . This allows displaying some additional data per point.

Note

There is no show_legend for a GraphConfiguration of a points graph. Instead you probably want to set the show_scale of the color_scale (and/or of the border and/or edges ). In addition, the color scale options of the edges must not be set, as the edges_colors of PointsGraphData is restricted to explicit colors.

MCGraphs.Renderers.PointsConfiguration — Type

@kwdef mutable struct PointsConfiguration
    color::Maybe{AbstractString} = nothing
    color_scale::ScaleConfiguration = ScaleConfiguration()
    color_palette::Maybe{Union{
        AbstractString,
        AbstractVector{<:Tuple{<:Real, <:AbstractString}},
        AbstractVector{<:Tuple{<:AbstractString, <:AbstractString}}
    }} = nothing
    size::Maybe{Real} = nothing
    size_scale::ScaleConfiguration = ScaleConfiguration()
    size_range::SizeRangeConfiguration = SizeRangeConfiguration()
end

Configure points in a graph. By default, the point color and size is chosen automatically (when this is applied to edges, the size is the width of the line). You can override this by specifying colors and/or sizes in the PointsGraphData . For color values, the color_palette is applied; this can be the name of a standard one, a vector of (value, color) tuples for a continuous (numeric value) scale or categorical (string value) scales. For sizes, the size_range is applied. The color_scale and size_scale configurations further control the color scales.

Grid

MCGraphs.Renderers.GridGraphData — Type

@kwdef mutable struct GridGraphData <: AbstractGraphData
    graph_title::Maybe{AbstractString} = nothing
    x_axis_title::Maybe{AbstractString} = nothing
    y_axis_title::Maybe{AbstractString} = nothing
    points_colors_title::Maybe{AbstractString} = nothing
    border_colors_title::Maybe{AbstractString} = nothing
    columns_names::Maybe{AbstractVector{<:AbstractString}} = nothing
    rows_names::Maybe{AbstractVector{<:AbstractString}} = nothing
    points_colors::Maybe{Union{AbstractMatrix{<:Union{AbstractString, <:Real}}}} = nothing
    points_sizes::Maybe{AbstractMatrix{<:Real}} = nothing
    points_hovers::Maybe{AbstractMatrix{<:AbstractString}} = nothing
    borders_colors::Maybe{Union{AbstractMatrix{<:Union{AbstractString, <:Real}}}} = nothing
    borders_sizes::Maybe{AbstractMatrix{<:Real}} = nothing
end

The data for a graph showing a grid of points (e.g. for correlations).

This is similar to a PointsGraphData , except that the data is given as a matrix instead of a vector, and no X and Y coordinates are given. Instead each matrix entry is plotted as a point at the matching grid location.

MCGraphs.Renderers.GridGraphConfiguration — Type

@kwdef mutable struct GridGraphConfiguration <: AbstractGraphConfiguration
    graph::GraphConfiguration = GraphConfiguration()
    points::PointsConfiguration = PointsConfiguration()
    border::PointsConfiguration = PointsConfiguration()
end

Configure a graph showing a grid of points (e.g. for correlations).

This displays a matrix of values using a point at each grid position. This is often used to show correlations between two sets of discrete variables. For large amounts of data, use a heatmap instead. An advantage of this over a heatmap is that you can control the color, size, and border color and size of the points to show more than one value per point. A disadvantage of this plot is that it only works for "small" amount of data. Also, you will need to manually tweak the graph size vs. the point sizes for best results, as Plotly's defaults aren't very good here.

Lines

MCGraphs.Renderers.LineGraphData — Type

@kwdef mutable struct LineGraphData <: AbstractGraphData
    graph_title::Maybe{AbstractString} = nothing
    x_axis_title::Maybe{AbstractString} = nothing
    y_axis_title::Maybe{AbstractString} = nothing
    points_xs::AbstractVector{<:Real}
    points_ys::AbstractVector{<:Real}
end

The data for a line graph.

By default, all the titles are empty. You can specify the overall graph_title as well as the x_axis_title and y_axis_title for the axes.

The points_xs and points_ys vectors must be of the same size. A line will be drawn through all the points, and the area under the line may be filled.

MCGraphs.Renderers.LinesGraphData — Type

@kwdef mutable struct LinesGraphData <: AbstractGraphData
    graph_title::Maybe{AbstractString} = nothing
    x_axis_title::Maybe{AbstractString} = nothing
    y_axis_title::Maybe{AbstractString} = nothing
    legend_title::Maybe{AbstractString} = nothing
    lines_xs::AbstractVector{AbstractVector{<:Real}}
    lines_ys::AbstractVector{AbstractVector{<:Real}}
    lines_names::Maybe{AbstractVector{<:AbstractString}} = nothing
    lines_colors::Maybe{AbstractVector{<:AbstractString}} = nothing
    lines_widths::Maybe{AbstractVector{<:Real}} = nothing
    lines_are_filled::Maybe{Union{AbstractVector{Bool}, BitVector}} = nothing
    lines_are_dashed::Maybe{AbstractVector{Bool}} = nothing
end

The data for multiple lines graph.

By default, all the titles are empty. You can specify the overall graph_title as well as the x_axis_title and y_axis_title for the axes.

The lines_xs and lines_ys vectors must be of the same size (one per line). For each line, its points xs and ys coordinate arrays must also be of the same size; a line will be drawn through all the points, and the area under the line may be filled. If stack_lines is specified in LinesGraphConfiguration , then the lines are specified in top-to-bottom order.

The lines_names , lines_colors , lines_widths , lines_are_filled and lines_are_dashed arrays must have the same number of entries (one per line). The lines_colors are restricted to explicit colors; therefore the color scale options of the line must not be used.

MCGraphs.Renderers.LineGraphConfiguration — Type

@kwdef mutable struct LineGraphConfiguration <: AbstractGraphConfiguration
    graph::GraphConfiguration = GraphConfiguration()
    x_axis::AxisConfiguration = AxisConfiguration()
    y_axis::AxisConfiguration = AxisConfiguration()
    line::LineConfiguration = LineConfiguration()
    vertical_bands::BandsConfiguration = BandsConfiguration()
    horizontal_bands::BandsConfiguration = BandsConfiguration()
end

Configure a graph for showing line plots.

MCGraphs.Renderers.LinesGraphConfiguration — Type

@kwdef mutable struct LinesGraphConfiguration <: AbstractGraphConfiguration
    graph::GraphConfiguration = GraphConfiguration()
    x_axis::AxisConfiguration = AxisConfiguration()
    y_axis::AxisConfiguration = AxisConfiguration()
    line::LineConfiguration = LineConfiguration()
    vertical_bands::BandsConfiguration = BandsConfiguration()
    horizontal_bands::BandsConfiguration = BandsConfiguration()
    show_legend::Bool = false
    data_stacking::Maybe{GraphDataStacking} = nothing
end

Configure a graph for showing multiple line plots. This allows using show_legend to display a legend of the different lines, and data_stacking to stack instead of overlay the lines. If data_stacking is specified, then is_filled is implied, regardless of what its actual setting is.

MCGraphs.Renderers.LineConfiguration — Type

@kwdef mutable struct LineConfiguration
    width::Maybe{Real} = 1
    is_filled::Bool = false
    is_dashed::Bool = false
    color::Maybe{AbstractString} = nothing
end

Configure a line in a graph.

By default, a solid line is shown; if is_dashed , the line will be dashed. If is_filled is set, the area below the line is filled. If the width is set to nothing , no line is shown (and is_filled must be set). By default, the color is chosen automatically.

CDFs

MCGraphs.Renderers.CdfGraphData — Type

@kwdef mutable struct CdfGraphData <: AbstractGraphData
    graph_title::Maybe{AbstractString} = nothing
    value_axis_title::Maybe{AbstractString} = nothing
    fraction_axis_title::Maybe{AbstractString} = nothing
    line_values::AbstractVector{<:Real}
end

The data for a CDF (Cumulative Distribution Function) graph.

By default, all the titles are empty. You can specify the overall graph_title as well as the value_axis_title and fraction_axis_title for the axes.

The order of the line_values does not matter.

MCGraphs.Renderers.CdfsGraphData — Type

@kwdef mutable struct CdfsGraphData <: AbstractGraphData
    graph_title::Maybe{AbstractString} = nothing
    value_axis_title::Maybe{AbstractString} = nothing
    fraction_axis_title::Maybe{AbstractString} = nothing
    lines_values::AbstractVector{<:AbstractVector{<:Real}}
    lines_names::Maybe{AbstractVector{<:AbstractString}} = nothing
    lines_colors::Maybe{AbstractVector{<:AbstractString}} = nothing
    lines_widths::Maybe{AbstractVector{<:Real}} = nothing
    lines_are_filled::Maybe{Union{AbstractVector{Bool}, BitVector}} = nothing
    lines_are_dashed::Maybe{AbstractVector{Bool}} = nothing
end

The data for multiple CDFs (Cumulative Distribution Functions) graph.

By default, all the titles are empty. You can specify the overall graph_title as well as the value_axis_title and fraction_axis_title for the axes.

The order of the entries inside each of the lines_values does not matter.

MCGraphs.Renderers.CdfsGraphConfiguration — Type

@kwdef mutable struct CdfsGraphConfiguration <: AbstractGraphConfiguration
    graph::GraphConfiguration = GraphConfiguration()
    value_axis::AxisConfiguration = AxisConfiguration()
    fraction_axis::AxisConfiguration = AxisConfiguration()
    line::LineConfiguration = LineConfiguration()
    values_orientation::ValuesOrientation = HorizontalValues
    cdf_direction::CdfDirection = CdfUpToValue
    value_bands::BandsConfiguration = BandsConfiguration()
    fraction_bands::BandsConfiguration = BandsConfiguration()
    show_legend::Bool = false
end

Configure a graph for showing multiple CDF (Cumulative Distribution Function) graph. This is the same as CdfGraphConfiguration with the addition of a show_legend field.

CDF graphs are internally converted to line graphs for rendering.

MCGraphs.Renderers.CdfGraphConfiguration — Type

@kwdef mutable struct CdfGraphConfiguration <: AbstractGraphConfiguration
    graph::GraphConfiguration = GraphConfiguration()
    value_axis::AxisConfiguration = AxisConfiguration()
    fraction_axis::AxisConfiguration = AxisConfiguration()
    line::LineConfiguration = LineConfiguration()
    values_orientation::ValuesOrientation = HorizontalValues
    cdf_direction::CdfDirection = CdfUpToValue
    value_bands::BandsConfiguration = BandsConfiguration()
    fraction_bands::BandsConfiguration = BandsConfiguration()
end

Configure a graph for showing a CDF (Cumulative Distribution Function) graph. By default, the X axis is used for the values and the Y axis for the fraction; this can be switched using the values_orientation . By default, the fraction is of the values up to each value; this can be switched using the cdf_direction .

By default, the fraction axis units are between 0 and 1; if show_percent , this is changed to between 0 and 100.

CDF graphs are internally converted to line graphs for rendering.

MCGraphs.Renderers.CdfDirection — Type

The direction of the CDF graph:

CdfUpToValue - Show the fraction of values up to each value.

CdfDownToValue - Show the fraction of values down to each value.

Bars

MCGraphs.Renderers.BarGraphData — Type

@kwdef mutable struct BarGraphData <: AbstractGraphData
    graph_title::Maybe{AbstractString} = nothing
    value_axis_title::Maybe{AbstractString} = nothing
    bar_axis_title::Maybe{AbstractString} = nothing
    bars_values::AbstractVector{<:Real}
    bars_names::Maybe{AbstractVector{<:AbstractString}} = nothing
    bars_colors::Maybe{AbstractVector{<:AbstractString}} = nothing
    bars_hovers::Maybe{AbstractVector{<:AbstractString}} = nothing
end

The data for a single bar (histogram) graph.

By default, all the titles are empty. You can specify the overall graph_title as well as the value_axis_title and bar_axis_title for the axes.

If specified, the bars_names and/or bars_colors and/or bars_hovers vectors must contain the same number of elements as the number of bars_values .

MCGraphs.Renderers.BarGraphConfiguration — Type

@kwdef mutable struct BarGraphConfiguration <: AbstractGraphConfiguration
    graph::GraphConfiguration = GraphConfiguration()
    value_axis::AxisConfiguration = AxisConfiguration()
    values_orientation::ValuesOrientation = VerticalValues
    bars_color::Maybe{AbstractString} = nothing
    bars_gap::Maybe{Real} = nothing
end

Configure a graph for showing a single bar (histogram) graph. The bars_color is chosen automatically. You can override it globally, or per-bar in the BarGraphData . By default, the X axis is used for the bars and the Y axis for the values; this can be switched using the values_orientation . The bars_gap is the fraction of white space between bars.

MCGraphs.Renderers.BarsGraphData — Type

@kwdef mutable struct BarsGraphData <: AbstractGraphData
    graph_title::Maybe{AbstractString} = nothing
    value_axis_title::Maybe{AbstractString} = nothing
    bar_axis_title::Maybe{AbstractString} = nothing
    legend_title::Maybe{AbstractString} = nothing
    series_values::AbstractString{<:AbstractVector{<:Real}}
    names::Maybe{AbstractVector{<:AbstractString}} = nothing
    colors::Maybe{AbstractVector{<:AbstractString}} = nothing
    hovers::Maybe{AbstractVector{<:AbstractString}} = nothing
    bars_names::Maybe{AbstractVector{<:AbstractString}} = nothing
end

The data for a multiple bars (histograms) graph.

By default, all the titles are empty. You can specify the overall graph_title as well as the value_axis_title , bar_axis_title for the axes, and the legend_title (if show_legend is set in BarsGraphConfiguration .

All the series_values vectors must be of the same size. If specified, the names vector must contain the same number of elements. If specified, the bars_names and/or colors and/or hovers vectors must contain the same number of elements as the number of series_values vectors.

MCGraphs.Renderers.BarsGraphConfiguration — Type

@kwdef mutable struct BarsGraphConfiguration <: AbstractGraphConfiguration
    graph::GraphConfiguration = GraphConfiguration()
    value_axis::AxisConfiguration = AxisConfiguration()
    color::Maybe{AbstractString} = nothing
    values_orientation::ValuesOrientation = VerticalValues
    bars_gap::Maybe{Real} = nothing
    show_legend::Bool = false
    data_stacking::Maybe{GraphDataStacking} = nothing
end

Configure a graph for showing multiple bars (histograms) graph. This is similar to BarGraphConfiguration , without the color field (which makes no sense when multiple series are shown), and with the addition of a show_legend and data_stacking fields. If data_stacking isn't specified then the different series are just grouped.

Distribution

MCGraphs.Renderers.DistributionGraphData — Type

@kwdef mutable struct DistributionGraphData <: AbstractGraphData
    graph_title::Maybe{AbstractString} = nothing
    value_axis_title::Maybe{AbstractString} = nothing
    trace_axis_title::Maybe{AbstractString} = nothing
    distribution_values::AbstractVector{<:Real}
    distribution_name::Maybe{AbstractString} = nothing
end

By default, all the titles are empty. You can specify the overall graph_title as well as the value_axis_title and the trace_axis_title . The optional distribution_name is used as the tick value for the distribution.

MCGraphs.Renderers.DistributionsGraphData — Type

@kwdef mutable struct DistributionsGraphData <: AbstractGraphData
    graph_title::Maybe{AbstractString} = nothing
    value_axis_title::Maybe{AbstractString} = nothing
    trace_axis_title::Maybe{AbstractString} = nothing
    legend_title::Maybe{AbstractString} = nothing
    distributions_values::AbstractVector{AbstractVector{<:Real}}
    distributions_names::Maybe{AbstractVector{<:AbstractString}} = nothing
    distributions_colors::Maybe{AbstractVector{<:AbstractString}} = nothing
end

The data for a multiple distributions graph. By default, all the titles are empty. You can specify the overall graph_title as well as the value_axis_title , the trace_axis_title and the legend_title (if show_legend is set). If specified, the distributions_names and/or the distributions_colors vectors must contain the same number of elements as the number of vectors in the distributions_values .

MCGraphs.Renderers.DistributionGraphConfiguration — Type

@kwdef mutable struct DistributionGraphConfiguration <: AbstractGraphConfiguration
    graph::GraphConfiguration = GraphConfiguration()
    distribution::DistributionConfiguration = DistributionConfiguration()
    value_axis::AxisConfiguration = AxisConfiguration()
end

Configure a graph for showing a distribution (with DistributionGraphData ) or several distributions (with DistributionsGraphData ).

The optional color will be chosen automatically if not specified. When showing multiple distributions, it is also possible to specify the color of each one in the DistributionsGraphData .

MCGraphs.Renderers.DistributionsGraphConfiguration — Type

@kwdef mutable struct DistributionsGraphConfiguration <: AbstractGraphConfiguration
    graph::GraphConfiguration = GraphConfiguration()
    distribution::DistributionConfiguration = DistributionConfiguration()
    value_axis::AxisConfiguration = AxisConfiguration()
    show_legend::Bool = false
    distributions_gap::Maybe{Real} = nothing
    overlay_distributions::Bool = false
end

Configure a graph for showing several distributions several distributions.

This is identical to DistributionGraphConfiguration with the addition of show_legend to show a legend. This is not set by default as it makes little sense unless overlay_distributions is also set. The distributions_gap is the fraction of white space between the distributions.

Note

Specifying a distributions_gap will end badly if using show_curve because Plotly.

MCGraphs.Renderers.DistributionConfiguration — Type

@kwdef mutable struct DistributionConfiguration
    values_orientation::ValuesOrientation = VerticalValues
    show_box::Bool = true
    show_violin::Bool = false
    show_curve::Bool = false
    show_outliers::Bool = false
    color::Maybe{AbstractString} = nothing
end

Configure the style of a distribution graph.

The values_orientation controls which axis is used for the values (the other axis is used for the density). By default the values are shown on the Y axis (vertical values).

If show_box , show a box graph.

If show_violin , show a violin graph.

If show_curve , show a density curve.

You can combine the above; however, a density curve is just the positive side of a violin graph, so you can't combine the two.

In addition to the (combination) of the above, if show_outliers , also show the extreme (outlier) points.

The color is chosen automatically by default. When showing multiple distributions, you can override it per each one in the DistributionsGraphData .

Common

MCGraphs.Renderers.GraphConfiguration — Type

@kwdef mutable struct GraphConfiguration
    width::Maybe{Int} = nothing
    height::Maybe{Int} = nothing
    template::AbstractString = "simple_white"
    show_grid::Bool = true
    show_ticks::Bool = true
end

Generic configuration that applies to any graph. Each complete AbstractGraphConfiguration contains a graph field of this type.

The optional width and height are in pixels, that is, 1/96 of an inch.

By default, show_grid and show_ticks are set.

The default template is "simple white" which is the cleanest. The `show grid and show_ticks` can be used to disable the grid and/or ticks for an even cleaner (but less informative) look.

MCGraphs.Renderers.AxisConfiguration — Type

@kwdef mutable struct AxisConfiguration
    minimum::Maybe{Real} = nothing
    maximum::Maybe{Real} = nothing
    log_regularization::Maybe{AbstractFloat} = nothing,
end

Generic configuration for a graph axis. Everything is optional; by default, the minimum and maximum are computed automatically from the data.

If log_regularization is set, it is added to the coordinate to avoid zero values, and the axis is shown in log (base

scale. To help with finer-grained ratios, each 10x step is broken to three ~2.15 steps (which is "close enough" to

2x for intuitive reading of the ratios).

MCGraphs.Renderers.ValuesOrientation — Type

The orientation of the values axis in a distribution or bars graph:

HorizontalValues - The values are the X axis

VerticalValues - The values are the Y axis (the default).

MCGraphs.Renderers.GraphDataStacking — Type

If stacking the data of multiple graphs, how:

StackDataValues - simply add the values on top of each other.

StackDataFractions - normalize the added values so their some is 1. The values must not be negative.

StackDataPercents - normalize the added values so their some is 100 (percent). The values must not be negative.

MCGraphs.Renderers.ScaleConfiguration — Type

@kwdef mutable struct ScaleConfiguration
    minimum::Maybe{Real} = nothing
    maximum::Maybe{Real} = nothing
    log_regularization::Maybe{AbstractFloat} = nothing
    reverse_scale::Bool = false
    show_scale::Bool = false
end

Configure a color (or sizes) scale. If show_scale is set, the color scale is displayed next to the graph. The rest of the fields are only relevant for continuous scales. By default, the minimum and maximum values are determined automatically by the data. Setting log_regularization converts the scale to a log scale, using the (non-negative) regularization to avoid log of zero values. If reverse_scale is set, the direction of the color scale is reversed.

MCGraphs.Renderers.SizeRangeConfiguration — Type

@kwdef mutable struct SizeRangeConfiguration
    smallest::Maybe{Real} = nothing
    largest::Maybe{Real} = nothing
end

Configure the range of sizes in pixels (1/96th of an inch) to map the sizes data into. If no bounds are given, and also the scale is linear, then we assume the sizes data is just the size in pixels. Otherwise, by default we use 2 pixels for the smallest size and make the largest size be 8 pixels larger than the smallest size.

MCGraphs.Renderers.BandsConfiguration — Type

@kwdef mutable struct BandsConfiguration
    low::BandConfiguration = BandConfiguration(is_dashed = true)
    middle::BandConfiguration = BandConfiguration()
    high::BandConfiguration = BandConfiguration(is_dashed = true)
end

Configure the partition of the graph up to three band regions. The low and high bands are for the "outer" regions (so their lines are at their border, dashed by default) and the middle band is for the "inner" region between them (so its line is inside it, solid by default).

MCGraphs.Renderers.BandConfiguration — Type

@kwdef mutable struct BandConfiguration
    offset::Maybe{Real} = nothing
    color::Maybe{AbstractString} = nothing
    width::Maybe{Real} = 1.0
    is_dashed::Bool = false
    is_filled::Bool = false
end

Configure a region of the graph defined by some band of values. This is the same as a LineConfiguration (for controlling the style of the line drawn for the band) with the addition of the offset of the line's position. We allow up to three bands in a complete BandsConfiguration . The low and high bands are defined as below and above their line's offset , and do not exist if the offset is not specified. The middle band is defined to be between these two lines (and therefore only exists if both are specified). Its offset defined a center line, if one is to be displayed, and is therefore optional.

Index

MCGraphs.Renderers

MCGraphs.Renderers.AbstractGraphConfiguration

MCGraphs.Renderers.AbstractGraphData

MCGraphs.Renderers.AxisConfiguration

MCGraphs.Renderers.BandConfiguration

MCGraphs.Renderers.BandsConfiguration

MCGraphs.Renderers.BarGraphConfiguration

MCGraphs.Renderers.BarGraphData

MCGraphs.Renderers.BarsGraphConfiguration

MCGraphs.Renderers.BarsGraphData

MCGraphs.Renderers.CdfDirection

MCGraphs.Renderers.CdfGraphConfiguration

MCGraphs.Renderers.CdfGraphData

MCGraphs.Renderers.CdfsGraphConfiguration

MCGraphs.Renderers.CdfsGraphData

MCGraphs.Renderers.DistributionConfiguration

MCGraphs.Renderers.DistributionGraphConfiguration

MCGraphs.Renderers.DistributionGraphData

MCGraphs.Renderers.DistributionsGraphConfiguration

MCGraphs.Renderers.DistributionsGraphData

MCGraphs.Renderers.Graph

MCGraphs.Renderers.GraphConfiguration

MCGraphs.Renderers.GraphDataStacking

MCGraphs.Renderers.GridGraphConfiguration

MCGraphs.Renderers.GridGraphData

MCGraphs.Renderers.LineConfiguration

MCGraphs.Renderers.LineGraphConfiguration

MCGraphs.Renderers.LineGraphData

MCGraphs.Renderers.LinesGraphConfiguration

MCGraphs.Renderers.LinesGraphData

MCGraphs.Renderers.PlotlyFigure

MCGraphs.Renderers.PointsConfiguration

MCGraphs.Renderers.PointsGraphConfiguration

MCGraphs.Renderers.PointsGraphData

MCGraphs.Renderers.ScaleConfiguration

MCGraphs.Renderers.SizeRangeConfiguration

MCGraphs.Renderers.ValuesOrientation

MCGraphs.Renderers.render